refactor(tests): modularize test utilities in `CalcMatchStatisticsTest`
- Extracted `TestingProject`, `TestingParseCallback`, and `TestingStatsConsumer` classes to standalone files for better reusability.
- Removed inline definitions and redundant `TestingProjectProperties` in `CalcMatchStatisticsTest`.
- Updated test setup with simplified constructors and externalized mock logic.
